requirements.txt
defusedxml==0.4.1
Django==1.9.1
django-allauth==0.24.1
django-appconf==1.0.1
django-autoslug==1.9.3
django-avatar==2.2.1
django-cities-light==3.2.0
django-compressor==2.0
django-crispy-forms==1.6.0
django-debug-toolbar==1.4
django-extensions==1.6.1
django-ratelimit-backend==1.0
django-sitetree==1.5.1
django-stronghold==0.2.7
django-widget-tweaks==1.4.1
html5lib==0.9999999
oauthlib==1.0.3
Pillow==3.1.0
progressbar2==3.6.0
psycopg2==2.6.1
python-memcached==1.57
python-slugify==1.2.0
python3-openid==3.0.9
pytz==2015.7
rcssmin==1.0.6
requests==2.9.1
requests-oauthlib==0.6.0
rjsmin==1.0.12
six==1.10.0
sqlparse==0.1.18
Unidecode==0.4.19
Werkzeug==0.11.3
wheel==0.24.0
